:construction: 1.21 port in progress 4
first launch-able commit
datagen fix
data components
effects

+++ src/main/java/org/mineplugin/locusazzurro/icaruswings/client/render/IWingsExpandable.java
@@ -2,5 +2,5 @@ package org.mineplugin.locusazzurro.icaruswings.client.render;
 
 public interface IWingsExpandable {
 	
-	float getExpansionFactor();
+	double getExpansionFactor();
 }
