+++ src/main/java/org/mineplugin/locusazzurro/icaruswings/client/render/layers/WingsLayer.java
@@ -82,8 +82,8 @@ public class WingsLayer<T extends LivingEntity, M extends EntityModel<T>> extend
 			matrixStackIn.pushPose();
 			matrixStackIn.translate(0.0D, 0.0D, 0.125D);
 			if((itemstack.getItem() instanceof IWingsExpandable wingsExpandable && entityLivingBaseIn.isFallFlying())) {
-				float scale = wingsExpandable.getExpansionFactor();
-				matrixStackIn.scale(scale, scale, 1.0f);
+				double scale = wingsExpandable.getExpansionFactor();
+				matrixStackIn.scale((float) scale, (float) scale, 1.0f);
 			}
 			this.getParentModel().copyPropertiesTo(this.elytraModel);
 			this.elytraModel.setupAnim(entityLivingBaseIn, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch);
